requirements.txt
# requirements.txt for pyberryplc-platform
# Do not install numpy via pip on Raspberry Pi. Use 'sudo apt install python3-numpy' instead.

annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
colorzero==2.0
fastapi==0.115.12
gpiozero==2.0.1
h11==0.16.0
idna==3.10
keyboard==0.13.5
# numpy==1.2.4
pigpio==1.78
pydantic==2.11.3
pydantic_core==2.33.1
pyserial==3.5
python-decouple==3.8
scipy==1.15.2
setuptools==79.0.1
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn[standard]==0.34.2
aiofiles==24.1.0
